pyproject.toml
[project]
name = "vidbot-extension"
version = "0.1.0"
description = "VidBot: Robotic affordance prediction from egocentric video"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "numpy>=1.26.3,<2.0",
    "transformers>=4.45.0",
    "lightning>=2.4.0",
    "open3d>=0.18.0",
    "opencv-python>=4.10.0",
    "omegaconf>=2.3.0",
    "easydict>=1.13",
    "flow-vis>=0.1",
    "numba>=0.60.0",
    "scikit-image>=0.24.0",
    "transformations>=2024.5.24",
    "ftfy>=6.2.0",
    "einops>=0.8.0",
    "gdown>=5.2.0",
    "scikit-learn>=1.5.0",
    "scipy>=1.14.0",
    "pandas>=2.2.0",
    "Pillow>=10.4.0",
    "matplotlib>=3.9.0",
    "pytest>=8.3.0",
    "pyqt6>=6.10.2",
    "moviepy>=2.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["algos", "models", "diffuser_utils", "vidbot_utils", "demos"]
